@@ -1,82 +1,91 @@
-import React, { Component } from 'react';
+// Dependancies
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import axios from 'axios';
 import Scream from '../components/scream/Scream';
 import StaticProfile from '../components/profile/StaticProfile';
+// Redux
+import { connect } from 'react-redux';
+import { getUserData } from '../redux/actions/dataActions';
+// MUI stuff
 import Grid from '@material-ui/core/Grid';
-
 import ScreamSkeleton from '../util/ScreamSkeleton';
 import ProfileSkeleton from '../util/ProfileSkeleton';
 
-import { connect } from 'react-redux';
-import { getUserData } from '../redux/actions/dataActions';
+const User = (props) => {
 
-class user extends Component {
-  state = {
-    profile: null,
-    screamIdParam: null
+  const { getUserData, data: { screams, loading } } = props;
+
+  const [profile, setProfile] = useState(null);
+  const [screamIdParam, setScreamIdParam] = useState(null);
+
+  useEffect(() => {
+
+    const { handle, screamId } = props.match.params;
+
+    if (screamId) setScreamIdParam(screamId);
+
+    getUserData(handle);
+    fetchUserDetails(handle);
+  }, [getUserData, props.match.params])
+
+  const fetchUserDetails = async (handle) => {
+    try {
+      let res = await axios.get(`/user/${handle}`);
+      setProfile(res.data.user);
+    } catch (err) {
+      console.log(err);
+    }
   };
-  componentDidMount() {
-    const handle = this.props.match.params.handle;
-    const screamId = this.props.match.params.screamId;
 
-    if (screamId) this.setState({ screamIdParam: screamId });
+  const screamsMarkup = loading ? (
 
-    this.props.getUserData(handle);
-    axios
-      .get(`/user/${handle}`)
-      .then((res) => {
-        this.setState({
-          profile: res.data.user
-        });
-      })
-      .catch((err) => console.log(err));
-  }
-  render() {
-    const { screams, loading } = this.props.data;
-    const { screamIdParam } = this.state;
+    <ScreamSkeleton />
+  ) : screams === null ? (
+    <p>No screams from this user</p>
+  ) : !screamIdParam ? (
+    screams.map((scream) => <Scream key={scream.screamId} scream={scream} />)
+  ) : (
+          screams.map(scream => {
+            if (scream.screamId === screamIdParam)
+              return <Scream key={scream.screamId} scream={scream} openDialog />
+            else return <Scream key={scream.screamId} scream={scream} />
+          })
+        );
 
-    const screamsMarkup = loading ? (
-      <ScreamSkeleton />
-    ) : screams === null ? (
-      <p>No screams from this user</p>
-    ) : !screamIdParam ? (
-      screams.map((scream) => <Scream key={scream.screamId} scream={scream} />)
-    ) : (
-      screams.map((scream) => {
-        if (scream.screamId !== screamIdParam)
-          return <Scream key={scream.screamId} scream={scream} />;
-        else return <Scream key={scream.screamId} scream={scream} openDialog />;
-      })
-    );
-
-    return (
-      <Grid container spacing={16}>
-        <Grid item sm={8} xs={12}>
-          {screamsMarkup}
-        </Grid>
-        <Grid item sm={4} xs={12}>
-          {this.state.profile === null ? (
-            <ProfileSkeleton />
-          ) : (
-            <StaticProfile profile={this.state.profile} />
-          )}
-        </Grid>
+  return (
+    <Grid container spacing={10}>
+      <Grid item sm={8} xs={12}>
+        {screamsMarkup}
       </Grid>
-    );
-  }
+      <Grid item sm={4} xs={12}>
+        {profile === null ? (
+          <ProfileSkeleton />
+        ) : (
+            <StaticProfile profile={profile} />
+          )}
+      </Grid>
+    </Grid>
+  );
 }
 
-user.propTypes = {
+// PropTypes  
+User.propTypes = {
   getUserData: PropTypes.func.isRequired,
   data: PropTypes.object.isRequired
 };
 
+// Pull state from Redux Store To Component
 const mapStateToProps = (state) => ({
   data: state.data
 });
 
+// Push Actions To Props
+const mapActionsToProps = {
+  getUserData
+};
+
 export default connect(
   mapStateToProps,
-  { getUserData }
-)(user);
+  mapActionsToProps
+)(User);
