@@ -1,36 +1,38 @@
-import React, { Component } from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import './App.css';
-import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
-import createMuiTheme from '@material-ui/core/styles/createMuiTheme';
+// Dependancies
+import React from 'react';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import jwtDecode from 'jwt-decode';
+import axios from 'axios';
+// Components
+import Navbar from './components/layout/Navbar';
+import AuthRoute from './util/AuthRoute';
+// Pages
+import Home from './pages/Home';
+import login from './pages/Login';
+import signup from './pages/Signup';
+import user from './pages/User';
 // Redux
 import { Provider } from 'react-redux';
 import store from './redux/store';
 import { SET_AUTHENTICATED } from './redux/types';
-import { logoutUser, getUserData } from './redux/actions/userActions';
-// Components
-import Navbar from './components/layout/Navbar';
-import themeObject from './util/theme';
-import AuthRoute from './util/AuthRoute';
-// Pages
-import home from './pages/home';
-import login from './pages/login';
-import signup from './pages/signup';
-import user from './pages/user';
+import { logOutUser, getUserData } from './redux/actions/userActions';
+// MUI stuff
+import './App.css';
+import { MuiThemeProvider } from '@material-ui/core/styles';
+import { createMuiTheme } from '@material-ui/core/styles';
+import { global__theme } from './util/theme';
 
-import axios from 'axios';
 
-const theme = createMuiTheme(themeObject);
+const theme = createMuiTheme(global__theme);
 
-axios.defaults.baseURL =
-  'https://europe-west1-socialape-d081e.cloudfunctions.net/api';
+axios.defaults.baseURL = 'https://europe-west6-screams-62f3b.cloudfunctions.net/api';
 
-const token = localStorage.FBIdToken;
+const token = localStorage.getItem('FBIdToken');
 if (token) {
   const decodedToken = jwtDecode(token);
+  // console.log(decodedToken); expired after 1 hour
   if (decodedToken.exp * 1000 < Date.now()) {
-    store.dispatch(logoutUser());
+    store.dispatch(logOutUser());
     window.location.href = '/login';
   } else {
     store.dispatch({ type: SET_AUTHENTICATED });
@@ -38,32 +40,27 @@ if (token) {
     store.dispatch(getUserData());
   }
 }
+const App = () => {
+  return (
+    <MuiThemeProvider theme={theme}>
+      <Provider store={store}>
 
-class App extends Component {
-  render() {
-    return (
-      <MuiThemeProvider theme={theme}>
-        <Provider store={store}>
-          <Router>
-            <Navbar />
-            <div className="container">
-              <Switch>
-                <Route exact path="/" component={home} />
-                <AuthRoute exact path="/login" component={login} />
-                <AuthRoute exact path="/signup" component={signup} />
-                <Route exact path="/users/:handle" component={user} />
-                <Route
-                  exact
-                  path="/users/:handle/scream/:screamId"
-                  component={user}
-                />
-              </Switch>
-            </div>
-          </Router>
-        </Provider>
-      </MuiThemeProvider>
-    );
-  }
+        <Router>
+          <Navbar />
+          <div className='container'>
+            <Switch>
+              <Route exact path='/' component={Home} />
+              <AuthRoute exact path='/login' component={login} />
+              <AuthRoute exact path='/signup' component={signup} />
+              <Route exact path='/users/:handle' component={user} />
+              <Route exact path='/users/:handle/scream/:screamId' component={user} />
+            </Switch>
+          </div>
+        </Router>
+
+      </Provider>
+    </MuiThemeProvider>
+  );
 }
 
 export default App;
