@@ -1,10 +1,12 @@
-import React, { Component, Fragment } from 'react';
-import { Link } from 'react-router-dom';
-import { connect } from 'react-redux';
+// Dependancies
+import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom';
 import MyButton from '../../util/MyButton';
-import PostScream from '../scream/PostScream';
+import PostScream from './../scream/PostScream';
 import Notifications from './Notifications';
+// Redux
+import { connect } from 'react-redux';
 // MUI stuff
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
@@ -12,46 +14,41 @@ import Button from '@material-ui/core/Button';
 // Icons
 import HomeIcon from '@material-ui/icons/Home';
 
-class Navbar extends Component {
-  render() {
-    const { authenticated } = this.props;
-    return (
-      <AppBar>
-        <Toolbar className="nav-container">
-          {authenticated ? (
+const Navbar = ({ authenticated }) => {
+
+  return (
+    <AppBar>
+      <Toolbar className='nav-container'>
+        {(authenticated) ? (
+          <Fragment>
+            <PostScream />
+            <Link to="/">
+              <MyButton tip="Home">
+                <HomeIcon />
+              </MyButton>
+            </Link>
+            <Notifications />
+          </Fragment>
+        ) : (
             <Fragment>
-              <PostScream />
-              <Link to="/">
-                <MyButton tip="Home">
-                  <HomeIcon />
-                </MyButton>
-              </Link>
-              <Notifications />
+              <Button color='inherit' component={Link} to='/login'>Login</Button>
+              <Button color='inherit' component={Link} to='/'>Home</Button>
+              <Button color='inherit' component={Link} to='/signup'>Signup</Button>
             </Fragment>
-          ) : (
-            <Fragment>
-              <Button color="inherit" component={Link} to="/login">
-                Login
-              </Button>
-              <Button color="inherit" component={Link} to="/">
-                Home
-              </Button>
-              <Button color="inherit" component={Link} to="/signup">
-                Signup
-              </Button>
-            </Fragment>
-          )}
-        </Toolbar>
-      </AppBar>
-    );
-  }
+          )
+        }
+      </Toolbar>
+    </AppBar >
+  )
 }
 
+// PropTypes
 Navbar.propTypes = {
   authenticated: PropTypes.bool.isRequired
-};
+}
 
-const mapStateToProps = (state) => ({
+// Pull state from Redux To Component
+const mapStateToProps = state => ({
   authenticated: state.user.authenticated
 });
 
