@@ -1,4 +1,5 @@
-import React, { Component, Fragment } from 'react';
+// Dependancies
+import React, { useState, Fragment } from 'react';
 import { Link } from 'react-router-dom';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
@@ -18,111 +19,118 @@ import ChatIcon from '@material-ui/icons/Chat';
 import { connect } from 'react-redux';
 import { markNotificationsRead } from '../../redux/actions/userActions';
 
-class Notifications extends Component {
-  state = {
-    anchorEl: null
-  };
-  handleOpen = (event) => {
-    this.setState({ anchorEl: event.target });
-  };
-  handleClose = () => {
-    this.setState({ anchorEl: null });
-  };
-  onMenuOpened = () => {
-    let unreadNotificationsIds = this.props.notifications
+const Notifications = ({ notifications, markNotificationsRead }) => {
+
+  const [anchorEl, setAnchorEl] = useState(null);
+
+  const handleOpen = (e) => setAnchorEl(e.currentTarget);
+
+  const handleClose = () => setAnchorEl(null);
+
+  const onMenuOpened = () => {
+    let unreadNotificationsIds = notifications
       .filter((not) => !not.read)
       .map((not) => not.notificationId);
-    this.props.markNotificationsRead(unreadNotificationsIds);
+    markNotificationsRead(unreadNotificationsIds);
   };
-  render() {
-    const notifications = this.props.notifications;
-    const anchorEl = this.state.anchorEl;
 
-    dayjs.extend(relativeTime);
+  dayjs.extend(relativeTime);
 
-    let notificationsIcon;
-    if (notifications && notifications.length > 0) {
-      notifications.filter((not) => not.read === false).length > 0
-        ? (notificationsIcon = (
-            <Badge
-              badgeContent={
-                notifications.filter((not) => not.read === false).length
-              }
-              color="secondary"
-            >
-              <NotificationsIcon />
-            </Badge>
-          ))
-        : (notificationsIcon = <NotificationsIcon />);
-    } else {
-      notificationsIcon = <NotificationsIcon />;
-    }
-    let notificationsMarkup =
-      notifications && notifications.length > 0 ? (
-        notifications.map((not) => {
-          const verb = not.type === 'like' ? 'liked' : 'commented on';
-          const time = dayjs(not.createdAt).fromNow();
-          const iconColor = not.read ? 'primary' : 'secondary';
-          const icon =
-            not.type === 'like' ? (
-              <FavoriteIcon color={iconColor} style={{ marginRight: 10 }} />
-            ) : (
+  // Notifications Icon Badge
+  let notificationsIcon;
+  if (notifications && notifications.length > 0) {
+    notifications.filter((not) => not.read === false).length > 0
+      ? (notificationsIcon = (
+        <Badge
+          badgeContent={
+            notifications.filter((not) => not.read === false).length
+          }
+          color="secondary"
+        >
+          <NotificationsIcon />
+        </Badge>
+      ))
+      : (notificationsIcon = <NotificationsIcon />);
+  } else {
+    notificationsIcon = <NotificationsIcon />;
+  }
+
+  // Notifications Markup
+  let notificationsMarkup =
+    notifications && notifications.length > 0 ? (
+      notifications.map((not, i) => {
+        const verb = not.type === 'like' ? 'liked' : 'commented on';
+        const time = dayjs(not.createdAt).fromNow();
+        const iconColor = not.read ? 'primary' : 'secondary';
+        const icon =
+          not.type === 'like' ? (
+            <FavoriteIcon color={iconColor} style={{ marginRight: 10 }} />
+          ) : (
               <ChatIcon color={iconColor} style={{ marginRight: 10 }} />
             );
 
-          return (
-            <MenuItem key={not.createdAt} onClick={this.handleClose}>
-              {icon}
-              <Typography
-                component={Link}
-                color="default"
-                variant="body1"
-                to={`/users/${not.recipient}/scream/${not.screamId}`}
-              >
-                {not.sender} {verb} your scream {time}
-              </Typography>
-            </MenuItem>
-          );
-        })
-      ) : (
-        <MenuItem onClick={this.handleClose}>
+        return (
+          <MenuItem key={i} onClick={handleClose}>
+            {icon}
+            <Typography
+              component={Link}
+              color="inherit"
+              variant="body1"
+              to={`/users/${not.recipient}/scream/${not.screamId}`}
+              target={"_blank"}
+            >
+              {not.sender} {verb} your scream {time}
+            </Typography>
+          </MenuItem>
+        );
+      })
+    ) : (
+        <MenuItem onClick={handleClose}>
           You have no notifications yet
         </MenuItem>
       );
-    return (
-      <Fragment>
-        <Tooltip placement="top" title="Notifications">
-          <IconButton
-            aria-owns={anchorEl ? 'simple-menu' : undefined}
-            aria-haspopup="true"
-            onClick={this.handleOpen}
-          >
-            {notificationsIcon}
-          </IconButton>
-        </Tooltip>
-        <Menu
-          anchorEl={anchorEl}
-          open={Boolean(anchorEl)}
-          onClose={this.handleClose}
-          onEntered={this.onMenuOpened}
+
+  return (
+    <Fragment>
+      <Tooltip placement="top" title="Notifications">
+        <IconButton
+          aria-owns={anchorEl ? 'simple-menu' : undefined}
+          aria-haspopup="true"
+          onClick={handleOpen}
         >
-          {notificationsMarkup}
-        </Menu>
-      </Fragment>
-    );
-  }
+          {notificationsIcon}
+        </IconButton>
+      </Tooltip>
+      <Menu
+        anchorEl={anchorEl}
+        open={Boolean(anchorEl)}
+        onClose={handleClose}
+        onEntered={onMenuOpened}
+      >
+        {notificationsMarkup}
+      </Menu>
+    </Fragment>
+  );
+
 }
 
+// PropTypes
 Notifications.propTypes = {
   markNotificationsRead: PropTypes.func.isRequired,
   notifications: PropTypes.array.isRequired
 };
 
+// Pull state from Redux Store To Component
 const mapStateToProps = (state) => ({
   notifications: state.user.notifications
 });
 
+// Push Actions To Props
+const mapActionsToProps = {
+  markNotificationsRead
+}
+
 export default connect(
   mapStateToProps,
-  { markNotificationsRead }
+  mapActionsToProps
 )(Notifications);
