@@ -1,113 +1,130 @@
-import React, { Component, Fragment } from 'react';
+// Dependancies
+import React, { Fragment, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import withStyles from '@material-ui/core/styles/withStyles';
 import { Link } from 'react-router-dom';
 import dayjs from 'dayjs';
-import EditDetails from './EditDetails';
 import MyButton from '../../util/MyButton';
-import ProfileSkeleton from '../../util/ProfileSkeleton';
+import EditDetails from './EditDetails';
+import ProfileSkeleton from './../../util/ProfileSkeleton';
+// Redux stuff
+import { connect } from 'react-redux';
+import { logOutUser, uploadImage } from '../../redux/actions/userActions';
+
 // MUI stuff
+import withStyles from '@material-ui/core/styles/withStyles';
 import Button from '@material-ui/core/Button';
-import Typography from '@material-ui/core/Typography';
-import MuiLink from '@material-ui/core/Link';
+import { profile__theme } from './../../util/theme';
 import Paper from '@material-ui/core/Paper';
+import MuiLink from '@material-ui/core/Link';
+import Typography from '@material-ui/core/Typography';
+import Alert from '@material-ui/lab/Alert';
 // Icons
 import LocationOn from '@material-ui/icons/LocationOn';
 import LinkIcon from '@material-ui/icons/Link';
 import CalendarToday from '@material-ui/icons/CalendarToday';
 import EditIcon from '@material-ui/icons/Edit';
 import KeyboardReturn from '@material-ui/icons/KeyboardReturn';
-//Redux
-import { connect } from 'react-redux';
-import { logoutUser, uploadImage } from '../../redux/actions/userActions';
 
-const styles = (theme) => ({
-  ...theme
-});
 
-class Profile extends Component {
-  handleImageChange = (event) => {
-    const image = event.target.files[0];
+const styles = { ...profile__theme };
+
+const Profile = (props) => {
+
+  const {
+    uploadImage,
+    logOutUser,
+    classes,
+    user: { credentials: { handle, createdAt, imageUrl, bio, website, location }, authenticated, loading }
+  } = props;
+
+  const [errors, setErrors] = useState({});
+
+  useEffect(() => {
+    if (props.UI && props.UI.errors)
+      setErrors(props.UI.errors);
+  }, [props]);
+
+  const handleImageChange = (e) => {
+    const image = e.target.files[0];
     const formData = new FormData();
     formData.append('image', image, image.name);
-    this.props.uploadImage(formData);
+    uploadImage(formData);
   };
-  handleEditPicture = () => {
+
+  const handleEditPicture = () => {
     const fileInput = document.getElementById('imageInput');
     fileInput.click();
   };
-  handleLogout = () => {
-    this.props.logoutUser();
-  };
-  render() {
-    const {
-      classes,
-      user: {
-        credentials: { handle, createdAt, imageUrl, bio, website, location },
-        loading,
-        authenticated
-      }
-    } = this.props;
 
-    let profileMarkup = !loading ? (
-      authenticated ? (
-        <Paper className={classes.paper}>
-          <div className={classes.profile}>
-            <div className="image-wrapper">
-              <img src={imageUrl} alt="profile" className="profile-image" />
-              <input
-                type="file"
-                id="imageInput"
-                hidden="hidden"
-                onChange={this.handleImageChange}
-              />
-              <MyButton
-                tip="Edit profile picture"
-                onClick={this.handleEditPicture}
-                btnClassName="button"
-              >
-                <EditIcon color="primary" />
-              </MyButton>
-            </div>
-            <hr />
-            <div className="profile-details">
-              <MuiLink
-                component={Link}
-                to={`/users/${handle}`}
-                color="primary"
-                variant="h5"
-              >
-                @{handle}
-              </MuiLink>
-              <hr />
-              {bio && <Typography variant="body2">{bio}</Typography>}
-              <hr />
-              {location && (
-                <Fragment>
-                  <LocationOn color="primary" /> <span>{location}</span>
-                  <hr />
-                </Fragment>
-              )}
-              {website && (
-                <Fragment>
-                  <LinkIcon color="primary" />
-                  <a href={website} target="_blank" rel="noopener noreferrer">
-                    {' '}
-                    {website}
-                  </a>
-                  <hr />
-                </Fragment>
-              )}
-              <CalendarToday color="primary" />{' '}
-              <span>Joined {dayjs(createdAt).format('MMM YYYY')}</span>
-            </div>
-            <MyButton tip="Logout" onClick={this.handleLogout}>
-              <KeyboardReturn color="primary" />
+  const handleLogout = () => logOutUser();
+
+  let profileMarkup = !loading ? (
+    authenticated ? (
+      <Paper className={classes.paper}>
+        <div className={classes.profile}>
+          <div className="image-wrapper">
+            {errors.error ? <Alert variant="outlined" severity="error">
+              {errors.error}
+            </Alert> : null}
+
+            {errors.image ? <Alert variant="outlined" severity="error">
+              {errors.image}
+            </Alert> : null}
+
+            <img src={imageUrl} alt="profile" className="profile-image" />
+            <input
+              type="file"
+              id="imageInput"
+              hidden="hidden"
+              onChange={handleImageChange}
+            />
+            <MyButton
+              tip="Edit profile picture"
+              onClick={handleEditPicture}
+              btnClassName="button"
+            >
+              <EditIcon color="primary" />
             </MyButton>
-            <EditDetails />
           </div>
-        </Paper>
-      ) : (
+          <hr />
+          <div className="profile-details">
+            <MuiLink
+              component={Link}
+              to={`/users/${handle}`}
+              color="primary"
+              variant="h6"
+            >
+              @{handle}
+            </MuiLink>
+            <hr />
+            {bio && <Typography variant="body2">{bio}</Typography>}
+            <hr />
+            {location && (
+              <Fragment>
+                <LocationOn color="primary" /> <span>{location}</span>
+                <hr />
+              </Fragment>
+            )}
+            {website && (
+              <Fragment>
+                <LinkIcon color="primary" />
+                <a href={website} target="_blank" rel="noopener noreferrer">
+                  {' '}
+                  {website}
+                </a>
+                <hr />
+              </Fragment>
+            )}
+            <CalendarToday color="primary" />{' '}
+            <span>Joined {dayjs(createdAt).format('MMM YYYY')}</span>
+          </div>
+          <MyButton tip="Logout" onClick={handleLogout}>
+            <KeyboardReturn color="primary" />
+          </MyButton>
+          <EditDetails />
+        </div>
+      </Paper>
+    ) : (
         <Paper className={classes.paper}>
           <Typography variant="body2" align="center">
             No profile found, please login again
@@ -132,28 +149,35 @@ class Profile extends Component {
           </div>
         </Paper>
       )
-    ) : (
+  ) : (
       <ProfileSkeleton />
     );
-
-    return profileMarkup;
-  }
+  return profileMarkup;
 }
 
+
+// PropTypes
+Profile.propTypes = {
+  classes: PropTypes.object.isRequired,
+  user: PropTypes.object.isRequired,
+  UI: PropTypes.object.isRequired,
+  logOutUser: PropTypes.func.isRequired,
+  uploadImage: PropTypes.func.isRequired
+}
+
+// Pull state from Redux Store To Component
 const mapStateToProps = (state) => ({
-  user: state.user
+  user: state.user,
+  UI: state.UI
 });
 
-const mapActionsToProps = { logoutUser, uploadImage };
-
-Profile.propTypes = {
-  logoutUser: PropTypes.func.isRequired,
-  uploadImage: PropTypes.func.isRequired,
-  user: PropTypes.object.isRequired,
-  classes: PropTypes.object.isRequired
-};
+// Push Actions To Props
+const mapActionsToProps = {
+  logOutUser,
+  uploadImage
+}
 
 export default connect(
   mapStateToProps,
-  mapActionsToProps
-)(withStyles(styles)(Profile));
+  mapActionsToProps)
+  (withStyles(styles)(Profile));
