@@ -1,15 +1,15 @@
-import React, { Component, Fragment } from 'react';
+// Dependancies
+import React, { Fragment, useState } from 'react';
 import withStyles from '@material-ui/core/styles/withStyles';
 import PropTypes from 'prop-types';
 import MyButton from '../../util/MyButton';
-
-// MUI Stuff
+// MUI stuff
 import Button from '@material-ui/core/Button';
 import Dialog from '@material-ui/core/Dialog';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import DialogActions from '@material-ui/core/DialogActions';
 import DeleteOutline from '@material-ui/icons/DeleteOutline';
-
+// Redux
 import { connect } from 'react-redux';
 import { deleteScream } from '../../redux/actions/dataActions';
 
@@ -21,62 +21,63 @@ const styles = {
   }
 };
 
-class DeleteScream extends Component {
-  state = {
-    open: false
-  };
-  handleOpen = () => {
-    this.setState({ open: true });
-  };
-  handleClose = () => {
-    this.setState({ open: false });
-  };
-  deleteScream = () => {
-    this.props.deleteScream(this.props.screamId);
-    this.setState({ open: false });
-  };
-  render() {
-    const { classes } = this.props;
+const DeleteScream = ({ classes, screamId, deleteScream }) => {
 
-    return (
-      <Fragment>
-        <MyButton
-          tip="Delete Scream"
-          onClick={this.handleOpen}
-          btnClassName={classes.deleteButton}
-        >
-          <DeleteOutline color="secondary" />
-        </MyButton>
-        <Dialog
-          open={this.state.open}
-          onClose={this.handleClose}
-          fullWidth
-          maxWidth="sm"
-        >
-          <DialogTitle>
-            Are you sure you want to delete this scream ?
+  const [open, setOpen] = useState(false)
+
+  const handleOpen = () => setOpen(true);
+
+  const handleClose = () => setOpen(false);
+
+  const handleDelete = () => {
+    deleteScream(screamId);
+    setOpen(false);
+  };
+
+  return (
+    <Fragment>
+      <MyButton
+        tip="Delete Scream"
+        onClick={handleOpen}
+        btnClassName={classes.deleteButton}
+      >
+        <DeleteOutline color="secondary" />
+      </MyButton>
+      <Dialog
+        open={open}
+        onClose={handleClose}
+        fullWidth
+        maxWidth="sm"
+      >
+        <DialogTitle>
+          Are you sure you want to delete this scream ?
           </DialogTitle>
-          <DialogActions>
-            <Button onClick={this.handleClose} color="primary">
-              Cancel
+        <DialogActions>
+          <Button onClick={handleClose} color="primary">
+            Cancel
             </Button>
-            <Button onClick={this.deleteScream} color="secondary">
-              Delete
+          <Button onClick={handleDelete} color="secondary">
+            Delete
             </Button>
-          </DialogActions>
-        </Dialog>
-      </Fragment>
-    );
-  }
+        </DialogActions>
+      </Dialog>
+    </Fragment>
+  );
 }
 
+// PropTypes
 DeleteScream.propTypes = {
   deleteScream: PropTypes.func.isRequired,
   classes: PropTypes.object.isRequired,
   screamId: PropTypes.string.isRequired
 };
 
+// Push Actions To Props
+const mapActionsToProps = {
+  deleteScream
+};
+
 export default connect(
   null,
-  { deleteScream }
+  mapActionsToProps
 )(withStyles(styles)(DeleteScream));
