@@ -1,109 +1,77 @@
-import React, { Component } from 'react';
-import withStyles from '@material-ui/core/styles/withStyles';
+// dependencies
+import React from 'react';
+import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
-import PropTypes from 'prop-types';
 import MyButton from '../../util/MyButton';
+import LikeButton from './LikeButton';
 import DeleteScream from './DeleteScream';
 import ScreamDialog from './ScreamDialog';
-import LikeButton from './LikeButton';
-// MUI Stuff
+// Redux
+import { connect } from 'react-redux';
+// MUI stuff
+import withStyles from '@material-ui/core/styles/withStyles';
+import { scream__theme } from './../../util/theme';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Typography from '@material-ui/core/Typography';
 // Icons
 import ChatIcon from '@material-ui/icons/Chat';
-// Redux
-import { connect } from 'react-redux';
 
-const styles = {
-  card: {
-    position: 'relative',
-    display: 'flex',
-    marginBottom: 20
-  },
-  image: {
-    minWidth: 200
-  },
-  content: {
-    padding: 25,
-    objectFit: 'cover'
-  }
-};
+const styles = { ...scream__theme };
 
-class Scream extends Component {
-  render() {
-    dayjs.extend(relativeTime);
-    const {
-      classes,
-      scream: {
-        body,
-        createdAt,
-        userImage,
-        userHandle,
-        screamId,
-        likeCount,
-        commentCount
-      },
-      user: {
-        authenticated,
-        credentials: { handle }
-      }
-    } = this.props;
+const Scream = (props) => {
 
-    const deleteButton =
-      authenticated && userHandle === handle ? (
-        <DeleteScream screamId={screamId} />
-      ) : null;
-    return (
-      <Card className={classes.card}>
-        <CardMedia
-          image={userImage}
-          title="Profile image"
-          className={classes.image}
-        />
-        <CardContent className={classes.content}>
-          <Typography
-            variant="h5"
-            component={Link}
-            to={`/users/${userHandle}`}
-            color="primary"
-          >
-            {userHandle}
-          </Typography>
-          {deleteButton}
-          <Typography variant="body2" color="textSecondary">
-            {dayjs(createdAt).fromNow()}
-          </Typography>
-          <Typography variant="body1">{body}</Typography>
-          <LikeButton screamId={screamId} />
-          <span>{likeCount} Likes</span>
-          <MyButton tip="comments">
-            <ChatIcon color="primary" />
-          </MyButton>
-          <span>{commentCount} comments</span>
-          <ScreamDialog
-            screamId={screamId}
-            userHandle={userHandle}
-            openDialog={this.props.openDialog}
-          />
-        </CardContent>
-      </Card>
-    );
-  }
+  const {
+    classes,
+    scream: { body, createdAt, userImage, userHandle, screamId, likeCount, commentCount },
+    user: { authenticated, credentials: { handle } },
+    openDialog
+  } = props;
+
+  const deleteButton =
+    authenticated && userHandle === handle ? (
+      <DeleteScream screamId={screamId} />
+    ) : null;
+
+  dayjs.extend(relativeTime);
+
+  return (
+    <Card className={classes.card} >
+      <CardMedia image={userImage} title='Profile Image' className={classes.image} />
+      <CardContent className={classes.content}>
+        <Typography variant='h5' color="primary" component={Link} to={`users/${userHandle}`}>{userHandle}</Typography>
+        <Typography variant='body2' color='textSecondary'><small color='textSecondary'>{`@${userHandle}`}</small></Typography>
+        {deleteButton}
+        <Typography variant='body2' color='textSecondary'>{dayjs(createdAt).fromNow()}</Typography>
+        <Typography variant='body1'>{body}</Typography>
+        <LikeButton screamId={screamId} />
+        <span>{likeCount} Likes</span>
+        <MyButton tip='comments'>
+          <ChatIcon color='primary' />
+        </MyButton>
+        <span>{commentCount} comments</span>
+        <ScreamDialog screamId={screamId} userHandle={userHandle} openDialog={openDialog} />
+      </CardContent>
+    </Card>
+  )
 }
 
+// PropTypes
 Scream.propTypes = {
   user: PropTypes.object.isRequired,
   scream: PropTypes.object.isRequired,
   classes: PropTypes.object.isRequired,
   openDialog: PropTypes.bool
-};
+}
 
-const mapStateToProps = (state) => ({
-  user: state.user
-});
+// Pull state from Redux Store To Component
+const mapStateToProps = (state) => {
+  return {
+    user: state.user
+  }
+};
 
 export default connect(mapStateToProps)(withStyles(styles)(Scream));
