@@ -1,163 +1,148 @@
-import React, { Component } from 'react';
-import withStyles from '@material-ui/core/styles/withStyles';
+// dependencies
+import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types';
-import AppIcon from '../images/icon.png';
 import { Link } from 'react-router-dom';
-
-// MUI Stuff
+import appIcon from './../images/icon.png';
+// MUI stuff
+import withStyles from '@material-ui/core/styles/withStyles';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import CircularProgress from '@material-ui/core/CircularProgress';
+import { signup__theme } from '../util/theme';
 // Redux stuff
 import { connect } from 'react-redux';
 import { signupUser } from '../redux/actions/userActions';
 
-const styles = (theme) => ({
-  ...theme
-});
+const styles = { ...signup__theme };
 
-class signup extends Component {
-  constructor() {
-    super();
-    this.state = {
-      email: '',
-      password: '',
-      confirmPassword: '',
-      handle: '',
-      errors: {}
-    };
-  }
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.UI.errors) {
-      this.setState({ errors: nextProps.UI.errors });
-    }
-  }
-  handleSubmit = (event) => {
-    event.preventDefault();
-    this.setState({
-      loading: true
-    });
+const Signup = (props) => {
+
+  const { classes, signupUser, UI: { loading } } = props;
+
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [handle, setHandle] = useState('');
+  const [errors, setErrors] = useState({});
+
+  useEffect(() => {
+    if (props.UI.errors)
+      setErrors(props.UI.errors);
+  }, [props]);
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
     const newUserData = {
-      email: this.state.email,
-      password: this.state.password,
-      confirmPassword: this.state.confirmPassword,
-      handle: this.state.handle
+      email,
+      password,
+      confirmPassword,
+      handle
     };
-    this.props.signupUser(newUserData, this.props.history);
-  };
-  handleChange = (event) => {
-    this.setState({
-      [event.target.name]: event.target.value
-    });
-  };
-  render() {
-    const {
-      classes,
-      UI: { loading }
-    } = this.props;
-    const { errors } = this.state;
-
-    return (
-      <Grid container className={classes.form}>
-        <Grid item sm />
-        <Grid item sm>
-          <img src={AppIcon} alt="monkey" className={classes.image} />
-          <Typography variant="h2" className={classes.pageTitle}>
-            SignUp
-          </Typography>
-          <form noValidate onSubmit={this.handleSubmit}>
-            <TextField
-              id="email"
-              name="email"
-              type="email"
-              label="Email"
-              className={classes.textField}
-              helperText={errors.email}
-              error={errors.email ? true : false}
-              value={this.state.email}
-              onChange={this.handleChange}
-              fullWidth
-            />
-            <TextField
-              id="password"
-              name="password"
-              type="password"
-              label="Password"
-              className={classes.textField}
-              helperText={errors.password}
-              error={errors.password ? true : false}
-              value={this.state.password}
-              onChange={this.handleChange}
-              fullWidth
-            />
-            <TextField
-              id="confirmPassword"
-              name="confirmPassword"
-              type="password"
-              label="Confirm Password"
-              className={classes.textField}
-              helperText={errors.confirmPassword}
-              error={errors.confirmPassword ? true : false}
-              value={this.state.confirmPassword}
-              onChange={this.handleChange}
-              fullWidth
-            />
-            <TextField
-              id="handle"
-              name="handle"
-              type="text"
-              label="Handle"
-              className={classes.textField}
-              helperText={errors.handle}
-              error={errors.handle ? true : false}
-              value={this.state.handle}
-              onChange={this.handleChange}
-              fullWidth
-            />
-            {errors.general && (
-              <Typography variant="body2" className={classes.customError}>
-                {errors.general}
-              </Typography>
-            )}
-            <Button
-              type="submit"
-              variant="contained"
-              color="primary"
-              className={classes.button}
-              disabled={loading}
-            >
-              SignUp
-              {loading && (
-                <CircularProgress size={30} className={classes.progress} />
-              )}
-            </Button>
-            <br />
-            <small>
-              Already have an account ? Login <Link to="/login">here</Link>
-            </small>
-          </form>
-        </Grid>
-        <Grid item sm />
-      </Grid>
-    );
+    signupUser(newUserData, props.history);
   }
+
+  return (
+    <Grid container className={classes.form}>
+      <Grid item sm />
+      <Grid item sm >
+        <img src={appIcon} alt='Screams Social Ape' className={classes.image} />
+        <Typography variant='h2' className={classes.pageTitle} >
+          Signup
+          </Typography>
+        <form noValidate onSubmit={handleSubmit}>
+          <TextField
+            id='email'
+            name='email'
+            type='email'
+            label='Email'
+            className={classes.textField}
+            helperText={errors.email}
+            error={errors.email ? true : false}
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            fullWidth />
+          <TextField
+            id='password'
+            name='password'
+            type='password'
+            label='Password'
+            className={classes.textField}
+            helperText={errors.password}
+            error={errors.password ? true : false} value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            fullWidth />
+          <TextField
+            id='confirmPassword'
+            name='confirmPassword'
+            type='password'
+            label='Confirm Password'
+            className={classes.textField}
+            helperText={errors.confirmPassword}
+            error={errors.confirmPassword ? true : false} value={confirmPassword}
+            onChange={(e) => setConfirmPassword(e.target.value)}
+            fullWidth />
+          <TextField
+            id='handle'
+            name='handle'
+            type='text'
+            label='handle'
+            className={classes.textField}
+            helperText={errors.handle}
+            error={errors.handle ? true : false} value={handle}
+            onChange={(e) => setHandle(e.target.value)}
+            fullWidth />
+
+          {errors.general && (
+            <Typography
+              variant='body2' className={classes.customeError}
+            >
+              {errors.general}
+            </Typography>
+          )}
+
+          <Button
+            variant='contained'
+            type='submit'
+            color='primary'
+            className={classes.button}
+            disabled={loading}
+          >
+            Sign up
+              {loading && (
+              <CircularProgress
+                className={classes.progress} size={30} />
+            )}
+          </Button><br />
+          <small>Already have an account ? log in <Link to='/login'>here</Link></small>
+        </form>
+      </Grid>
+      <Grid item sm />
+    </Grid>
+  );
 }
 
-signup.propTypes = {
+// PropTypes
+Signup.protoTypes = {
   classes: PropTypes.object.isRequired,
+  signupUser: PropTypes.func.isRequired,
   user: PropTypes.object.isRequired,
-  UI: PropTypes.object.isRequired,
-  signupUser: PropTypes.func.isRequired
+  UI: PropTypes.object.isRequired
 };
 
+// Pull state from Redux Store To Component
 const mapStateToProps = (state) => ({
   user: state.user,
   UI: state.UI
 });
 
+// Push Actions To Props
+const mapActionsToProps = {
+  signupUser
+};
+
 export default connect(
   mapStateToProps,
-  { signupUser }
-)(withStyles(styles)(signup));
+  mapActionsToProps)
+  (withStyles(styles)(Signup));
