@@ -1,29 +1,33 @@
+// validation 
 import {
-  SET_SCREAMS,
-  LOADING_DATA,
-  LIKE_SCREAM,
-  UNLIKE_SCREAM,
-  DELETE_SCREAM,
-  SET_ERRORS,
-  POST_SCREAM,
-  CLEAR_ERRORS,
-  LOADING_UI,
-  SET_SCREAM,
-  STOP_LOADING_UI,
-  SUBMIT_COMMENT
-} from '../types';
+  validateCommentData,
+  validateScreamData
+} from './../util/validators';
+// types
+import {
+  SET_SCREAMS, LOADING_DATA, LIKE_SCREAM,
+  UNLIKE_SCREAM, DELETE_SCREAM, POST_SCREAM,
+  CLEAR_ERRORS, LOADING_UI, STOP_LOADING_UI,
+  SET_SCREAM, SUBMIT_COMMENT, SET_ERRORS,
+  LOADING_COMMENT_UI
+} from './../types';
 import axios from 'axios';
 
-// Get all screams
+// clear all errors
+export const clearErrors = () => (dispatch) => {
+  dispatch({ type: CLEAR_ERRORS });
+};
+
+// get all screams
 export const getScreams = () => (dispatch) => {
   dispatch({ type: LOADING_DATA });
   axios
     .get('/screams')
-    .then((res) => {
+    .then(res => {
       dispatch({
         type: SET_SCREAMS,
         payload: res.data
-      });
+      })
     })
     .catch((err) => {
       dispatch({
@@ -32,6 +36,77 @@ export const getScreams = () => (dispatch) => {
       });
     });
 };
+
+// like scream
+export const likeScream = (screamId) => (dispatch) => {
+  axios
+    .get(`/scream/${screamId}/like`)
+    .then(res => {
+      dispatch({
+        type: LIKE_SCREAM,
+        payload: res.data
+      })
+    })
+    .catch((err) => console.log(err));
+};
+
+// unlike scream
+export const unLikeScream = (screamId) => (dispatch) => {
+  axios
+    .get(`/scream/${screamId}/unlike`)
+    .then(res => {
+      dispatch({
+        type: UNLIKE_SCREAM,
+        payload: res.data
+      })
+    })
+    .catch((err) => console.log(err));
+};
+
+// delete scream
+export const deleteScream = (screamId) => (dispatch) => {
+  axios
+    .delete(`/scream/${screamId}`)
+    .then(res => {
+      dispatch({
+        type: DELETE_SCREAM,
+        payload: screamId
+      })
+    })
+    .catch((err) => console.log(err));
+};
+
+// psot a scream
+export const postScream = (newScream) => (dispatch) => {
+  const { valid, errors } = validateScreamData(newScream);
+  dispatch({ type: LOADING_UI });
+
+  if (valid) {
+    axios
+    .post('/scream', newScream)
+    .then(res => {
+      dispatch({
+        type: POST_SCREAM,
+        payload: res.data
+      });
+      dispatch({ type: CLEAR_ERRORS });
+    })
+    .catch((err) => {
+      dispatch({
+        type: SET_ERRORS,
+        payload: err.response.data
+      });
+    });
+  } else {
+    dispatch({
+      type: SET_ERRORS,
+      payload: errors
+    })
+  }
+
+}
+
+// get scream details
 export const getScream = (screamId) => (dispatch) => {
   dispatch({ type: LOADING_UI });
   axios
@@ -45,81 +120,45 @@ export const getScream = (screamId) => (dispatch) => {
     })
     .catch((err) => console.log(err));
 };
-// Post a scream
-export const postScream = (newScream) => (dispatch) => {
-  dispatch({ type: LOADING_UI });
-  axios
-    .post('/scream', newScream)
-    .then((res) => {
-      dispatch({
-        type: POST_SCREAM,
-        payload: res.data
-      });
-      dispatch(clearErrors());
-    })
-    .catch((err) => {
-      dispatch({
-        type: SET_ERRORS,
-        payload: err.response.data
-      });
-    });
-};
-// Like a scream
-export const likeScream = (screamId) => (dispatch) => {
-  axios
-    .get(`/scream/${screamId}/like`)
-    .then((res) => {
-      dispatch({
-        type: LIKE_SCREAM,
-        payload: res.data
-      });
-    })
-    .catch((err) => console.log(err));
-};
-// Unlike a scream
-export const unlikeScream = (screamId) => (dispatch) => {
-  axios
-    .get(`/scream/${screamId}/unlike`)
-    .then((res) => {
-      dispatch({
-        type: UNLIKE_SCREAM,
-        payload: res.data
-      });
-    })
-    .catch((err) => console.log(err));
-};
-// Submit a comment
+
+// comment on scream
 export const submitComment = (screamId, commentData) => (dispatch) => {
-  axios
-    .post(`/scream/${screamId}/comment`, commentData)
-    .then((res) => {
-      dispatch({
-        type: SUBMIT_COMMENT,
-        payload: res.data
-      });
-      dispatch(clearErrors());
-    })
-    .catch((err) => {
-      dispatch({
-        type: SET_ERRORS,
-        payload: err.response.data
-      });
+  const { errors, valid } = validateCommentData(commentData);
+
+  dispatch(clearErrors());
+
+  if (!valid) {
+    dispatch({ type: LOADING_COMMENT_UI });
+    dispatch({
+      type: SET_ERRORS,
+      payload: errors
     });
-};
-export const deleteScream = (screamId) => (dispatch) => {
-  axios
-    .delete(`/scream/${screamId}`)
-    .then(() => {
-      dispatch({ type: DELETE_SCREAM, payload: screamId });
-    })
-    .catch((err) => console.log(err));
+  } else {
+    axios
+      .post(`/scream/${screamId}/comment`, commentData)
+      .then((res) => {
+        dispatch({
+          type: SUBMIT_COMMENT,
+          payload: res.data
+        });
+        dispatch(clearErrors());
+      })
+      .catch((err) => {
+        dispatch({
+          type: SET_ERRORS,
+          payload: err.response.data
+        });
+      });
+  }
+
 };
 
+// get all data about on user
 export const getUserData = (userHandle) => (dispatch) => {
   dispatch({ type: LOADING_DATA });
   axios
     .get(`/user/${userHandle}`)
-    .then((res) => {
+    .then(res => {
       dispatch({
         type: SET_SCREAMS,
         payload: res.data.screams
@@ -131,8 +170,4 @@ export const getUserData = (userHandle) => (dispatch) => {
         payload: null
       });
     });
-};
-
-export const clearErrors = () => (dispatch) => {
-  dispatch({ type: CLEAR_ERRORS });
 };
