@@ -1,16 +1,12 @@
-import {
-  SET_ERRORS,
-  CLEAR_ERRORS,
-  LOADING_UI,
-  STOP_LOADING_UI
-} from '../types';
+// Types
+import { SET_ERRORS, CLEAR_ERRORS, LOADING_UI, STOP_LOADING_UI, LOADING_COMMENT_UI } from './../types';
 
 const initialState = {
   loading: false,
   errors: null
 };
 
-export default function(state = initialState, action) {
+export default function (state = initialState, action) {
   switch (action.type) {
     case SET_ERRORS:
       return {
@@ -22,18 +18,20 @@ export default function(state = initialState, action) {
       return {
         ...state,
         loading: false,
-        errors: null
+        errors: action.payload
       };
     case LOADING_UI:
       return {
         ...state,
-        loading: true
+        loading: SVGComponentTransferFunctionElement
       };
+
     case STOP_LOADING_UI:
       return {
         ...state,
         loading: false
-      };
+      }
+
     default:
       return state;
   }
