@@ -1,12 +1,9 @@
+// Types
 import {
-  SET_USER,
-  SET_AUTHENTICATED,
-  SET_UNAUTHENTICATED,
-  LOADING_USER,
-  LIKE_SCREAM,
-  UNLIKE_SCREAM,
-  MARK_NOTIFICATIONS_READ
-} from '../types';
+  SET_USER, LIKE_SCREAM, UNLIKE_SCREAM, LOADING_USER,
+  SET_AUTHENTICATED, SET_UNAUTHENTICATED, MARK_NOTIFICATIONS_READ,
+  STOP_LOADING_USER
+} from './../types';
 
 const initialState = {
   authenticated: false,
@@ -16,13 +13,14 @@ const initialState = {
   notifications: []
 };
 
-export default function(state = initialState, action) {
+export default function (state = initialState, action) {
   switch (action.type) {
     case SET_AUTHENTICATED:
       return {
         ...state,
         authenticated: true
       };
+
     case SET_UNAUTHENTICATED:
       return initialState;
     case SET_USER:
@@ -30,12 +28,20 @@ export default function(state = initialState, action) {
         authenticated: true,
         loading: false,
         ...action.payload
-      };
+      }
+
     case LOADING_USER:
       return {
         ...state,
-        loading: true
-      };
+        loading: true,
+      }
+
+    case STOP_LOADING_USER:
+      return {
+        ...state,
+        loading: false,
+      }
+
     case LIKE_SCREAM:
       return {
         ...state,
@@ -46,19 +52,20 @@ export default function(state = initialState, action) {
             screamId: action.payload.screamId
           }
         ]
-      };
+      }
+
     case UNLIKE_SCREAM:
       return {
         ...state,
-        likes: state.likes.filter(
-          (like) => like.screamId !== action.payload.screamId
-        )
-      };
+        likes: state.likes.filter(like => like.screamId !== action.payload.screamId)
+      }
+
     case MARK_NOTIFICATIONS_READ:
-      state.notifications.forEach((not) => (not.read = true));
+      state.notifications.forEach((notification) => (notification.read = true));
       return {
         ...state
-      };
+      }
+
     default:
       return state;
   }
