@@ -1,42 +1,26 @@
+// Dependancies
 import React from 'react';
 import PropTypes from 'prop-types';
-import withStyles from '@material-ui/core/styles/withStyles';
-import NoImg from '../images/no-img.png';
-// MUI
+import NoImg from './images/No-Image-Thumb.jpg';
+// MUI stuff
 import Paper from '@material-ui/core/Paper';
+import withStyles from '@material-ui/core/styles/withStyles';
+import { profile__theme, profile__skeleton__theme } from './theme';
 // Icons
 import LocationOn from '@material-ui/icons/LocationOn';
 import LinkIcon from '@material-ui/icons/Link';
 import CalendarToday from '@material-ui/icons/CalendarToday';
 
-const styles = (theme) => ({
-  ...theme,
-  handle: {
-    height: 20,
-    backgroundColor: theme.palette.primary.main,
-    width: 60,
-    margin: '0 auto 7px auto'
-  },
-  fullLine: {
-    height: 15,
-    backgroundColor: 'rgba(0,0,0,0.6)',
-    width: '100%',
-    marginBottom: 10
-  },
-  halfLine: {
-    height: 15,
-    backgroundColor: 'rgba(0,0,0,0.6)',
-    width: '50%',
-    marginBottom: 10
-  }
-});
+const styles = {
+  profile__theme,
+  ...profile__skeleton__theme
+};
 
-const ProfileSkeleton = (props) => {
-  const { classes } = props;
+const ProfileSkeleton = ({ classes }) => {
   return (
     <Paper className={classes.paper}>
       <div className={classes.profile}>
-        <div className="image-wrapper">
+        <div className={"image-wrapper"}>
           <img src={NoImg} alt="profile" className="profile-image" />
         </div>
         <hr />
@@ -57,6 +41,7 @@ const ProfileSkeleton = (props) => {
   );
 };
 
+// PropTypes
 ProfileSkeleton.propTypes = {
   classes: PropTypes.object.isRequired
 };
