@@ -1,56 +1,19 @@
+// Dependancies
 import React, { Fragment } from 'react';
-import NoImg from '../images/no-img.png';
+import NoImg from './images/No-Image-Thumb.jpg';
 import PropTypes from 'prop-types';
-// MUI
+// MUI stuff
+import { scream__skeleton__theme } from './theme';
 import Card from '@material-ui/core/Card';
 import CardMedia from '@material-ui/core/CardMedia';
 import CardContent from '@material-ui/core/CardContent';
-
 import withStyles from '@material-ui/core/styles/withStyles';
 
-const styles = (theme) => ({
-  ...theme,
-  card: {
-    display: 'flex',
-    marginBottom: 20
-  },
-  cardContent: {
-    width: '100%',
-    flexDirection: 'column',
-    padding: 25
-  },
-  cover: {
-    minWidth: 200,
-    objectFit: 'cover'
-  },
-  handle: {
-    width: 60,
-    height: 18,
-    backgroundColor: theme.palette.primary.main,
-    marginBottom: 7
-  },
-  date: {
-    height: 14,
-    width: 100,
-    backgroundColor: 'rgba(0,0,0, 0.3)',
-    marginBottom: 10
-  },
-  fullLine: {
-    height: 15,
-    width: '90%',
-    backgroundColor: 'rgba(0,0,0, 0.6)',
-    marginBottom: 10
-  },
-  halfLine: {
-    height: 15,
-    width: '50%',
-    backgroundColor: 'rgba(0,0,0, 0.6)',
-    marginBottom: 10
-  }
-});
+const styles = {
+  ...scream__skeleton__theme
+};
 
-const ScreamSkeleton = (props) => {
-  const { classes } = props;
+const ScreamSkeleton = ({ classes }) => {
 
   const content = Array.from({ length: 5 }).map((item, index) => (
     <Card className={classes.card} key={index}>
@@ -68,6 +31,7 @@ const ScreamSkeleton = (props) => {
   return <Fragment>{content}</Fragment>;
 };
 
+// PropTypes
 ScreamSkeleton.propTypes = {
   classes: PropTypes.object.isRequired
 };
