@@ -1,29 +1,37 @@
-export default {
+export const global__theme = {
   palette: {
     primary: {
-      light: '#33c9dc',
-      main: '#00bcd4',
-      dark: '#008394',
-      contrastText: '#fff'
+      main: '#441e51',
+      contrastText: '#fff',
     },
     secondary: {
-      light: '#ff6333',
-      main: '#ff3d00',
-      dark: '#b22a00',
-      contrastText: '#fff'
-    }
+      main: '#e4056f',
+      contrastText: '#fff',
+    },
   },
   typography: {
-    useNextVariants: true
+    useNextVariant: true
   },
+  invisibleSeparator: {
+    border: 'none',
+    margin: 4
+  },
+  visibleSeparator: {
+    width: '100%',
+    borderBottom: '1px solid rgba(0, 0, 0, 0.1)',
+    marginBottom: 20
+  }
+}
+
+export const signup__theme = {
   form: {
     textAlign: 'center'
   },
   image: {
-    margin: '20px auto 20px auto'
+    margin: '10px auto 20px auto'
   },
   pageTitle: {
-    margin: '10px auto 10px auto'
+    margin: '10px auto 5px auto'
   },
   textField: {
     margin: '10px auto 10px auto'
@@ -32,23 +40,44 @@ export default {
     marginTop: 20,
     position: 'relative'
   },
-  customError: {
+  customeError: {
     color: 'red',
     fontSize: '0.8rem',
-    marginTop: 10
+    marginTop: '10px'
   },
   progress: {
     position: 'absolute'
+  }
+}
+
+export const login__theme = {
+  form: {
+    textAlign: 'center'
   },
-  invisibleSeparator: {
-    border: 'none',
-    margin: 4
+  image: {
+    margin: '10px auto 20px auto'
   },
-  visibleSeparator: {
-    width: '100%',
-    borderBottom: '1px solid rgba(0,0,0,0.1)',
-    marginBottom: 20
+  pageTitle: {
+    margin: '10px auto 5px auto'
   },
+  textField: {
+    margin: '10px auto 10px auto'
+  },
+  button: {
+    marginTop: 20,
+    position: 'relative'
+  },
+  customeError: {
+    color: 'red',
+    fontSize: '0.8rem',
+    marginTop: '10px'
+  },
+  progress: {
+    position: 'absolute'
+  }
+}
+
+export const profile__theme = {
   paper: {
     padding: 20
   },
@@ -65,7 +94,7 @@ export default {
     '& .profile-image': {
       width: 200,
       height: 200,
-      objectFit: 'cover',
+      // objectFit: 'cover',
       maxWidth: '100%',
       borderRadius: '50%'
     },
@@ -75,7 +104,8 @@ export default {
         verticalAlign: 'middle'
       },
       '& a': {
-        color: '#00bcd4'
+        color: '#441e51',
+        fontWeight: 'bold'
       }
     },
     '& hr': {
@@ -94,4 +124,179 @@ export default {
       margin: '20px 10px'
     }
   }
-};
+}
+
+export const scream__theme = {
+  card: {
+    position: 'relative',
+    display: 'flex',
+    marginBottom: 20,
+  },
+  image: {
+    minWidth: 200
+  },
+  content: {
+    padding: 25,
+    objectFit: 'cover',
+    width: '100%'
+  }
+}
+
+export const post__scream__theme = {
+  submitButton: {
+    position: 'relative',
+    float: 'right',
+    marginTop: 10
+  },
+  progressSpinner: {
+    position: 'absolute'
+  },
+  closeButton: {
+    position: 'absolute',
+    left: '91%',
+    top: '6%'
+  }
+}
+
+export const scream__dialog__theme = {
+  profileImage: {
+    maxWidth: 200,
+    height: 200,
+    borderRadius: '50%',
+    // objectFit: 'cover'
+  },
+  dialogContent: {
+    padding: 20
+  },
+  closeButton: {
+    position: 'absolute',
+    left: '90%'
+  },
+  expandButton: {
+    position: 'absolute',
+    left: '90%'
+  },
+  spinnerDiv: {
+    textAlign: 'center',
+    marginTop: 50,
+    marginBottom: 50
+  }
+}
+
+export const comments__theme = {
+  commentImage: {
+    maxWidth: '100%',
+    height: '100px',
+    objectFit: 'cover',
+    borderRadius: '50%'
+  },
+  commentData: {
+    marginLeft: 20
+  }
+}
+
+export const scream__skeleton__theme = {
+  card: {
+    display: 'flex',
+    marginBottom: 20
+  },
+  cardContent: {
+    width: '100%',
+    flexDirection: 'column',
+    padding: 25
+  },
+  cover: {
+    minWidth: 200,
+    objectFit: 'cover'
+  },
+  handle: {
+    width: 60,
+    height: 18,
+    backgroundColor: global__theme.palette.primary.main,
+    marginBottom: 7
+  },
+  date: {
+    height: 14,
+    width: 100,
+    backgroundColor: 'rgba(0,0,0, 0.3)',
+    marginBottom: 10
+  },
+  fullLine: {
+    height: 15,
+    width: '90%',
+    backgroundColor: 'rgba(0,0,0, 0.6)',
+    marginBottom: 10
+  },
+  halfLine: {
+    height: 15,
+    width: '50%',
+    backgroundColor: 'rgba(0,0,0, 0.6)',
+    marginBottom: 10
+  }
+}
+
+export const profile__skeleton__theme = {
+  card: {
+    display: 'flex',
+    marginBottom: 20
+  },
+  cardContent: {
+    width: '100%',
+    flexDirection: 'column',
+    padding: 25
+  },
+  cover: {
+    minWidth: 200,
+    objectFit: 'cover'
+  },
+  handle: {
+    height: 20,
+    backgroundColor: global__theme.palette.primary.main,
+    width: 60,
+    margin: '0 auto 7px auto'
+  },
+  fullLine: {
+    height: 15,
+    backgroundColor: 'rgba(0,0,0,0.6)',
+    width: '100%',
+    marginBottom: 10
+  },
+  halfLine: {
+    height: 15,
+    backgroundColor: 'rgba(0,0,0,0.6)',
+    width: '50%',
+    marginBottom: 10
+  },
+  paper: {
+    padding: 20
+  },
+  profile: {
+    '& .image-wrapper': {
+      textAlign: 'center',
+      position: 'relative',
+      '& button': {
+        position: 'absolute',
+        top: '80%',
+        left: '70%'
+      }
+    },
+    '& .profile-image': {
+      width: 200,
+      height: 200,
+      // objectFit: 'cover',
+      maxWidth: '100%',
+      borderRadius: '50%'
+    },
+    '& .profile-details': {
+      textAlign: 'center',
+      '& span, svg': {
+        verticalAlign: 'middle'
+      },
+      '& svg.button': {
+        '&:hover': {
+          cursor: 'pointer'
+        }
+      }
+    }
+  }
+}
